Add button to ChooseServicePage
